[Help] Refactor App to live in index.js
