+++ src/hmda-help/index.css
@@ -1,6 +1,7 @@
 html {
   font-size: 10px;
 }
+
 body {
   margin: 0;
   padding: 0;
@@ -8,7 +9,36 @@ body {
     Arial, sans-serif;
   font-size: 1.6rem;
 }
+
 #root {
   max-width: 1040px;
   margin: 0 auto;
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: 50% 50%;
+}
+
+nav {
+  margin-top: 1.5em;
+  text-align: right;
+}
+
+nav a {
+  margin-right: 2em;
+}
+
+label {
+  display: block;
+  font-weight: bold;
+  margin-top: 3rem;
+  max-width: 46rem;
+}
+
+.unstyled-list {
+  margin-top: 0;
+  margin-bottom: 0;
+  list-style-type: none;
+  padding-left: 0;
 }
